less warns on msvc2013

+++ core/patterndetection/chessBoardAssembler.cpp
@@ -10,7 +10,7 @@ ChessBoardAssembler::ChessBoardAssembler(ChessBoardAssemblerParams params) : Che
 void ChessBoardAssembler::assembleBoards(std::vector<OrientedCorner> &corners_, std::vector<std::vector<std::vector<corecvs::Vector2dd>>> &boards_)
 {
     corners = corners_;
-    int N = corners.size();
+    int N = (int)corners.size();
 
     corecvs::parallelable_for(0, N, ParallelBoardExpander(this), true);
 
@@ -185,7 +185,7 @@ bool ChessBoardAssembler::BoardExpander::initBoard(int seed)
             ssq[i] += v * v;
         }
         int N;
-        s[i] /= (N = dv[i].size());
+        s[i] /= (N = (int)dv[i].size());
         ssq[i] = std::sqrt(ssq[i] / N - s[i] * s[i]);
 
         if (ssq[i] / s[i] > varThreshold)
@@ -203,7 +203,7 @@ bool ChessBoardAssembler::BoardExpander::getNearest(int from_id, corecvs::Vector
     auto& from = corners[from_id];
 
     double best_dist = 1e100;
-    int best_id = -1, N = corners.size();
+    int best_id = -1, N = (int)corners.size();
 
     for (int i = 0; i < N; ++i)
     {
@@ -381,8 +381,8 @@ bool ChessBoardAssembler::BoardExpander::assignNearest(std::vector<corecvs::Vect
         if (!usedCorners[i])
             unused.push_back(i);
 
-    int N = unused.size();
-    int M = prediction.size();
+    int N = (int)unused.size();
+    int M = (int)prediction.size();
     std::vector<int> assigned(M);
     assignment.resize(M);
